agent in device management panel
